move tabs to parent folder (components)
